producer/requirements.txt
# Producer requirements
celery==5.3.4
python-dotenv==1.0.0
requests==2.32.5
minio==7.2.16
